pyproject.toml
[tool.poetry]
name = "voyeur"
version = "0.1.0"
description = "A Django project for metrics collection"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14"
drf-yasg = "^1.21"
pymongo = {extras = ["srv"], version = "^4.6"}
python-dotenv = "^1.0"
django-cors-headers = "^4.7.0"
stomp-py = "^8.1.0"
websocket-client = "^1.7.0"
redis = "^6.2.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
sockjs-tornado = "^1.0.7"
tornado = "^6.4"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"